Compile.py now uses all stages, but ugly.
1) I copied some of the old compile code into compile.py to get it
limping along, but it's really ugly. From the selector on, it doesn't
conform to the interface.
I'm doing this so that I at least have a functional compiler. This will
make testing easier as I rewrite the selector class, and convert it into
a visitor. Once it's converted, I'll make it conform to the interface.