PrettyPrint works up until closure.
1) I added getChildNodes() methods for the nodes in p1ast.py. This gets
the simple test program up to the closure stage, but crashes when trying
to print the closure ast.
2) One problem was that the closure ast is actually inside a list. I
added code to handle that in prettyPrint().
3) Now the built in ast Function node is crashing when called
getChildNodes(). It says "NoneType not iterable." I'm not sure if this
is a bug in the compiler.ast module, or if the Function nodes just
aren't well formed. Maybe I set an attribute to None that shouldn't have
been.